common: Improve RAFChannelProvider
Use /dev/null instead of temporary files whenever possible.

+++ junixsocket-common/src/main/java/org/newsclub/net/unix/RAFChannelProvider.java
@@ -31,6 +31,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
  * @author Christian Kohlschütter
  */
 final class RAFChannelProvider extends RandomAccessFile implements FileDescriptorAccess {
+  private static final File DEV_NULL = new File("/dev/null");
+
   private final File tempPath;
   private final FileDescriptor fdObj;
   private final FileDescriptor rafFdOrig = new FileDescriptor();
@@ -41,21 +43,28 @@ final class RAFChannelProvider extends RandomAccessFile implements FileDescripto
   }
 
   private RAFChannelProvider(FileDescriptor fdObj, File tempPath) throws IOException {
-    super(tempPath, "rw");
-    this.tempPath = tempPath;
-    if (!tempPath.delete() && tempPath.exists()) {
-      if (!tempPath.delete()) {
-        // we tried our best (looking at you, Windows)
-      }
-      tempPath = new File(tempPath.getParentFile(), "jux-" + UUID.randomUUID().toString()
-          + ".sock");
-      if (tempPath.exists()) {
-        throw new IOException("Could not create a temporary path: " + tempPath);
-      }
-    }
-    tempPath.deleteOnExit();
+    this(fdObj, tempPath, true);
+  }
 
+  private RAFChannelProvider(FileDescriptor fdObj, File tempPath, boolean delete)
+      throws IOException {
+    super(tempPath, "rw");
     NativeUnixSocket.ensureSupported();
+
+    this.tempPath = tempPath;
+    if (delete) {
+      if (!tempPath.delete() && tempPath.exists()) {
+        if (!tempPath.delete()) {
+          // we tried our best (looking at you, Windows)
+        }
+        tempPath = new File(tempPath.getParentFile(), "jux-" + UUID.randomUUID().toString()
+            + ".sock");
+        if (tempPath.exists()) {
+          throw new IOException("Could not create a temporary path: " + tempPath);
+        }
+      }
+      tempPath.deleteOnExit();
+    }
 
     this.fdObj = fdObj;
 
@@ -85,6 +94,13 @@ final class RAFChannelProvider extends RandomAccessFile implements FileDescripto
 
   @SuppressWarnings("resource")
   private static FileChannel getFileChannel0(FileDescriptor fd) throws IOException {
+    if (DEV_NULL.canRead() && DEV_NULL.canWrite()) {
+      try {
+        return new RAFChannelProvider(fd, DEV_NULL, false).getChannel();
+      } catch (IOException e) {
+        // ignore
+      }
+    }
     return new RAFChannelProvider(fd).getChannel();
   }
 }
